requirements.txt
ipython==9.1.0
nbautoexport==0.5.2
pyyaml==6.0.2
matplotlib==3.10.1
numpy==2.1.3
pandas==2.2.3
dask[dataframe]==2025.3.0
torch==2.6.0
transformers==4.50.3
datasets[audio]==3.5.0